Fix warnings. Enhance readabilaty.

+++ src/optimization.h
@@ -8,7 +8,7 @@ const float R_MAX_SQUARED = 10.f;
 
 // Finds iteration of dropout (<MAX_ITERATION) for every pixel of the screen.
 int *
-opt_set_pixels(int *pixels, int x_0, int y_0, float scale);
+opt_set_pixels(int *pixels, float x_0, float y_0, float scale);
 // Same as previous; uses AVX optimization.
 void
 opt_set_pix_avx(int *pixels, float x_0, float y_0, float scale);
